feat: about-section.js module
